+++ src/extensions/ui/folder-tree/api/index.ts
@@ -1,2 +1,8 @@
-export { fetchFolder, fetchFolderPath, fetchFolders, moveFolder } from './folders';
+export {
+  clearFolderCache,
+  fetchFolder,
+  fetchFolderPath,
+  fetchFolders,
+  moveFolder,
+} from './folders';
 export { fetchWorkflowProjectId, moveWorkflow } from './workflows';